Move EventTypes and links into distinct files
