+++ lib/userEvents/reportOnUserEvents.js
@@ -1,29 +1,15 @@
+
 const describeTime = require('../describeTime');
+
+const EventTypes = require('./EventTypes');
 
 const renderUserEvent =  require('./renderHelpers').renderUserEvent;
 const renderIndex = require('./renderHelpers').renderIndex;
 
-const EventTypes = {
-  CreateEvent: 'CreateEvent',
-  DeleteEvent: 'DeleteEvent',
-  PullRequestEvent: 'PullRequestEvent',
-  PullRequestReviewCommentEvent: 'PullRequestReviewCommentEvent',
-  PushEvent: 'PushEvent',
-  WatchEvent: 'WatchEvent'
-};
+const buildLinkToViewCreatedItem = require('./linkBuilders').buildLinkToViewCreatedItem;
+const buildLinkToViewPushEventCommits = require('./linkBuilders').buildLinkToViewPushEventCommits;
 
 const removeRefsHeads = (ref) => ref.substring('refs/heads/'.length);
-
-const link = (href, text) => `<a href="${href}">${text}</a>`;
-
-const hrefForRepo = (server, repo) => `https://${server}/${repo}`;
-
-const hrefForBranch = (server, repo, branch) => `${hrefForRepo(server, repo)}/tree/${branch}`;
-
-const hrefForCommitsFromPushEvent = (server, repo, pushEventPayload) => {
-  const { before, head } = pushEventPayload;
-  return `${hrefForRepo(server, repo)}/compare/${before}...${head}`;
-};
 
 const buildUserEvent = (data, event, server) => {
   const eventLinkBuilder = eventLinks[event.type] || defaultLinkBuilder;
@@ -39,7 +25,6 @@ const buildUserEvent = (data, event, server) => {
   return userEvent;
 };
 
-// See also: https://developer.github.com/v3/activity/events/types/
 const eventDescriptionBuilders = {
   [EventTypes.PushEvent]: (event, server) => {
     const customData = {
@@ -68,22 +53,10 @@ const eventDescriptionBuilders = {
 
 const eventLinks = {
   [EventTypes.PushEvent]: (event, server) => {
-    const hrefToViewCommits = hrefForCommitsFromPushEvent(server, event.repo.name, event.payload);
-    const linkText = `the ${event.payload.size} commit(s)`;
-    return [ link(hrefToViewCommits, linkText) ];
+    return [ buildLinkToViewPushEventCommits(event, server) ];
   },
   [EventTypes.CreateEvent]: (event, server) => {
-    let hrefToViewCreatedThing;
-    let linkText;
-    const payload = event.payload;
-    if (payload.ref_type === 'repository') {
-      hrefToViewCreatedThing = hrefForRepo(server, event.repo.name, event.repo.name);
-      linkText = 'the new repo';
-    } else {
-      hrefToViewCreatedThing = hrefForBranch(server, event.repo.name, payload.ref);
-      linkText = `the new ${payload.ref_type}`;
-    }
-    return [ link(hrefToViewCreatedThing, linkText) ];
+    return [ buildLinkToViewCreatedItem(event, server) ];
   }
 };
